Implementation General Performance table
